Decompose LeftPanel.tsx into focused hooks and panel components
LeftPanel.tsx had grown to 1633 lines handling geometry import, meshing,
BC/load management, solve orchestration and results display in one file
(issue #165). Split it along the boundaries proposed there:

- hooks/useGeometry.ts: STEP/IGES import state + parse_step worker call
- hooks/useMesh.ts: mesh sizing params, meshing/log state + volume_mesh
  worker call and post-Netgen worker reset
- hooks/useSolver.ts: pre-flight checks, solve worker call + result
  storage (incl. vonMises), results-mode hand-off
- components/panel/GeometryPanel.tsx: import cards + material section
- components/panel/MeshPanel.tsx: mesh controls + meshing log
- components/panel/BoundaryConditionsPanel.tsx: BC/load groups, face
  picking, inline value editors
- components/panel/SolvePanel.tsx, ResultsPanel.tsx, PanelNav.tsx
- LeftPanel.tsx: thin orchestrator mapping mode -> panel (24 lines)

Pure code motion — no behavior change; each hook owns its slice of the
worker protocol so the solver path is testable without a full mount.

closes #165
