+++ web/src/components/panel/LeftPanel.tsx
@@ -1,1631 +1,24 @@
 // SPDX-FileCopyrightText: 2026 Michael Kofler
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
-import { useState, useRef, useEffect, type ChangeEvent } from "react";
-import {
-  useModelStore,
-  RESULT_TYPES,
-  loadKind,
-  loadComponents,
-} from "../../store/modelStore";
-import type {
-  AppMode,
-  LoadKind,
-  Material,
-  NamedBcGroup,
-  NamedLoadGroup,
-  Node,
-  Element,
-  ResultType,
-} from "../../store/modelStore";
-import { fmt } from "../../lib/modelDisplay";
-import {
-  computeResultRange,
-  resultFieldSymbol,
-  resultUnit,
-} from "../../lib/resultField";
-import {
-  sendToWorker,
-  setLogCallback,
-  resetWorker,
-} from "../../workers/sharedWorker";
+import { useModelStore } from "../../store/modelStore";
+import { PanelNav } from "./PanelNav";
+import { GeometryPanel } from "./GeometryPanel";
+import { BoundaryConditionsPanel } from "./BoundaryConditionsPanel";
+import { SolvePanel } from "./SolvePanel";
+import { ResultsPanel } from "./ResultsPanel";
 import styles from "./LeftPanel.module.css";
 
-// ── Geometry mode ─────────────────────────────────────────────────────────────
-
-function MaterialForm({
-  mat,
-  onSave,
-  onCancel,
-}: {
-  mat?: Material;
-  onSave(v: Omit<Material, "id">): void;
-  onCancel(): void;
-}) {
-  const [name, setName] = useState(mat?.name ?? "Material");
-  const [young, setYoung] = useState(String(mat?.young ?? 210000));
-  const [poisson, setPoisson] = useState(String(mat?.poisson ?? 0.3));
-  const [density, setDensity] = useState(String(mat?.density ?? 7.85e-9));
-  const [error, setError] = useState<string | null>(null);
-
-  function handleSave() {
-    const e = parseFloat(young);
-    const nu = parseFloat(poisson);
-    const rho = parseFloat(density);
-
-    if (!isFinite(e) || e <= 0) {
-      setError("Young's modulus must be a positive number");
-      return;
-    }
-    // ν must lie in (-1, 0.5); ν = 0 is physically valid (e.g. auxetic foams,
-    // cork) and must not be silently replaced by the steel default.
-    if (!isFinite(nu) || nu <= -1 || nu >= 0.5) {
-      setError("Poisson's ratio must be in the range (-1, 0.5)");
-      return;
-    }
-    if (!isFinite(rho) || rho < 0) {
-      setError("Density must be a non-negative number");
-      return;
-    }
-    onSave({ name: name || "Material", young: e, poisson: nu, density: rho });
-  }
-
-  return (
-    <div className={styles.inlineForm}>
-      {error && (
-        <div className={styles.errorBanner} data-testid="material-error">
-          <span>{error}</span>
-          <button onClick={() => setError(null)}>×</button>
-        </div>
-      )}
-      <div className={styles.formRow}>
-        <span className={styles.formLabel}>Name</span>
-        <input
-          className={styles.formInput}
-          value={name}
-          onChange={(e) => setName(e.target.value)}
-        />
-      </div>
-      <div className={styles.formRow}>
-        <span className={styles.formLabel}>E (MPa)</span>
-        <input
-          className={styles.formInput}
-          type="number"
-          value={young}
-          step="1000"
-          onChange={(e) => setYoung(e.target.value)}
-        />
-      </div>
-      <div className={styles.formRow}>
-        <span className={styles.formLabel}>ν</span>
-        <input
-          className={styles.formInput}
-          type="number"
-          value={poisson}
-          step="0.01"
-          onChange={(e) => setPoisson(e.target.value)}
-        />
-      </div>
-      <div className={styles.formRow}>
-        <span className={styles.formLabel}>ρ (t/mm³)</span>
-        <input
-          className={styles.formInput}
-          type="number"
-          value={density}
-          step="1e-9"
-          onChange={(e) => setDensity(e.target.value)}
-          title="Stored with the material but not used by the current static solver (reserved for future gravity / inertial loads)."
-        />
-      </div>
-      <p className={styles.formNote}>
-        Not used by the current static solver — reserved for future gravity /
-        inertial loads.
-      </p>
-      <div className={styles.formBtns}>
-        <button className={styles.cancelBtn} onClick={onCancel}>
-          Cancel
-        </button>
-        <button className={styles.primaryBtn} onClick={handleSave}>
-          Save
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function GeometryPanel() {
-  const nodes = useModelStore((s) => s.nodes);
-  const setStepSurface = useModelStore((s) => s.setStepSurface);
-  const stepImportError = useModelStore((s) => s.stepImportError);
-  const setStepImportError = useModelStore((s) => s.setStepImportError);
-  const isRunning = useModelStore((s) => s.isRunning);
-  const setRunning = useModelStore((s) => s.setRunning);
-  const materials = useModelStore((s) => s.materials);
-  const updateMaterial = useModelStore((s) => s.updateMaterial);
-  const deleteMaterial = useModelStore((s) => s.deleteMaterial);
-  const elementOrder = useModelStore((s) => s.elementOrder);
-  const setElementOrder = useModelStore((s) => s.setElementOrder);
-  const stepSurface = useModelStore((s) => s.stepSurface);
-  const stepBytes = useModelStore((s) => s.stepBytes);
-  const setStepBytes = useModelStore((s) => s.setStepBytes);
-  const geometryFormat = useModelStore((s) => s.geometryFormat);
-  const setGeometryFormat = useModelStore((s) => s.setGeometryFormat);
-  const isMeshing = useModelStore((s) => s.isMeshing);
-  const setMeshing = useModelStore((s) => s.setMeshing);
-  const applyMeshResult = useModelStore((s) => s.applyMeshResult);
-
-  const [editingMatId, setEditingMatId] = useState<number | null>(null);
-  const [isImportingStep, setIsImportingStep] = useState(false);
-  const [maxElementSize, setMaxElementSize] = useState(20);
-  // Floor for curvature-driven refinement; 0 lets Netgen refine fillets without
-  // limit, which can produce >10x more elements than the max size suggests.
-  const [minElementSize, setMinElementSize] = useState(2);
-  const [meshError, setMeshError] = useState<string | null>(null);
-  const [logs, setLogs] = useState<string[]>([]);
-  const [logsOpen, setLogsOpen] = useState(true);
-  const logsEndRef = useRef<HTMLDivElement | null>(null);
-
-  const stepRef = useRef<HTMLInputElement | null>(null);
-  const igesRef = useRef<HTMLInputElement | null>(null);
-
-  useEffect(() => {
-    setLogCallback((msg) => {
-      console.log("[mesh-log]", msg);
-      setLogs((prev) => [...prev, msg]);
-    });
-    return () => setLogCallback(null);
-  }, []);
-
-  useEffect(() => {
-    logsEndRef.current?.scrollIntoView({ behavior: "smooth" });
-  }, [logs]);
-
-  async function handleCadFile(e: ChangeEvent<HTMLInputElement>) {
-    const file = e.target.files?.[0];
-    if (!file) return;
-    e.target.value = "";
-    // Derive the reader from the extension so the same handler serves both the
-    // STEP and IGES file inputs (and a file picked through the "wrong" one).
-    const ext = file.name.split(".").pop()?.toLowerCase();
-    const format = ext === "igs" || ext === "iges" ? "iges" : "step";
-    const label = format === "iges" ? "IGES" : "STEP";
-    setStepImportError(null);
-    setIsImportingStep(true);
-    setRunning(true);
-    const bytes = new Uint8Array(await file.arrayBuffer());
-    sendToWorker<{
-      points: [number, number, number][];
-      triangles: [number, number, number][];
-    }>("parse_step", { bytes, format })
-      .then(({ points, triangles }) => {
-        if (points.length === 0) setStepImportError("No geometry found.");
-        else {
-          // Retain the raw bytes + format so the geometry can be reloaded for a
-          // re-mesh (the worker is reset after each mesh and loses the shape).
-          setStepBytes(bytes);
-          setGeometryFormat(format);
-          setStepSurface({ points, triangles });
-        }
-      })
-      .catch((err) =>
-        setStepImportError(err.message ?? `${label} import failed`),
-      )
-      .finally(() => {
-        setIsImportingStep(false);
-        setRunning(false);
-      });
-  }
-
-  async function handleVolMesh() {
-    if (!stepSurface) return;
-    if (!stepBytes) {
-      setMeshError(
-        "Cannot mesh: the original STEP file is no longer available (e.g. after loading a saved analysis). Re-import the STEP file to mesh.",
-      );
-      return;
-    }
-    setMeshing(true);
-    setLogs([]);
-    try {
-      const {
-        nodes: n,
-        elements: e,
-        surfaceTriangles,
-        surfaceFaceIds,
-      } = await sendToWorker<{
-        nodes: Node[];
-        elements: Element[];
-        surfaceTriangles: [number, number, number][] | null;
-        surfaceFaceIds: number[] | null;
-      }>("volume_mesh", {
-        bytes: stepBytes,
-        format: geometryFormat,
-        maxElementSize,
-        minElementSize,
-      });
-      applyMeshResult(
-        n,
-        e,
-        "STEP Volume Mesh",
-        surfaceTriangles,
-        surfaceFaceIds,
-      );
-      // Netgen's Ng_Init() installs global C++ state that contaminates the WASM
-      // runtime for subsequent MFEM solves.  Resetting the worker here gives the
-      // solve a clean module instance, preventing an infinite loop on first call.
-      resetWorker();
-    } catch (err) {
-      console.error("[meshing] volume mesh failed:", err);
-      const detail = err instanceof Error ? err.message : String(err);
-      setMeshError(`Volume meshing failed: ${detail}`);
-    } finally {
-      setMeshing(false);
-    }
-  }
-
-  // Mesh stats (node/element counts, solver-ready state) live in the status
-  // bar at the bottom — the panel only carries the meshing controls.
-  const hasMesh = nodes.length > 0;
-  const showLogs = logs.length > 0 || isMeshing;
-
-  return (
-    <div className={styles.panel}>
-      <div className={styles.tabContent}>
-        {/* ── Inputs ─────────────────────────────────────────── */}
-        <>
-          <input
-            ref={(el) => {
-              stepRef.current = el;
-            }}
-            type="file"
-            accept=".stp,.step"
-            style={{ display: "none" }}
-            onChange={handleCadFile}
-          />
-          <input
-            ref={(el) => {
-              igesRef.current = el;
-            }}
-            type="file"
-            accept=".igs,.iges"
-            style={{ display: "none" }}
-            onChange={handleCadFile}
-          />
-
-          <div className={styles.cardGrid}>
-            <button
-              className={styles.importCard}
-              disabled={isImportingStep || isRunning}
-              onClick={() => stepRef.current?.click()}
-            >
-              <svg className={styles.cardIcon} viewBox="0 0 20 20" fill="none">
-                <rect
-                  x="3"
-                  y="2"
-                  width="14"
-                  height="16"
-                  rx="2"
-                  stroke="currentColor"
-                  strokeWidth="1.4"
-                />
-                <path
-                  d="M7 7h6M7 10h6M7 13h4"
-                  stroke="currentColor"
-                  strokeWidth="1.4"
-                  strokeLinecap="round"
-                />
-              </svg>
-              <span className={styles.cardTitle}>
-                {isImportingStep ? "Importing…" : "Import STEP"}
-              </span>
-              <span className={styles.cardSub}>.step / .stp</span>
-            </button>
-
-            <button
-              className={styles.importCard}
-              disabled={isImportingStep || isRunning}
-              onClick={() => igesRef.current?.click()}
-            >
-              <svg className={styles.cardIcon} viewBox="0 0 20 20" fill="none">
-                <path
-                  d="M10 2l2.4 5H18l-4.2 3.1 1.6 5L10 12.2 4.6 15.1l1.6-5L2 7h5.6z"
-                  stroke="currentColor"
-                  strokeWidth="1.4"
-                />
-              </svg>
-              <span className={styles.cardTitle}>
-                {isImportingStep ? "Importing…" : "Import IGES"}
-              </span>
-              <span className={styles.cardSub}>.igs / .iges</span>
-            </button>
-          </div>
-
-          {stepImportError && (
-            <div
-              data-testid="step-error"
-              style={{ color: "#dc2626", fontSize: 12, padding: "4px 0" }}
-            >
-              {stepImportError}
-            </div>
-          )}
-        </>
-
-        {/* ── Mesh ─────────────────────────────────────────────── */}
-        <>
-          {meshError && (
-            <div className={styles.errorBanner} data-testid="meshing-error">
-              <span>{meshError}</span>
-              <button onClick={() => setMeshError(null)}>×</button>
-            </div>
-          )}
-
-          {stepSurface ? (
-            <>
-              <div className={styles.sectionLabel}>Mesh controls</div>
-              <div className={styles.formRow}>
-                <span className={styles.formLabel}>Max element size</span>
-                <input
-                  className={styles.formInput}
-                  type="number"
-                  min={0.5}
-                  max={500}
-                  step={0.5}
-                  value={maxElementSize}
-                  disabled={isMeshing}
-                  onChange={(e) =>
-                    setMaxElementSize(Math.max(0.5, Number(e.target.value)))
-                  }
-                />
-                <span className={styles.toleranceUnit}>mm</span>
-              </div>
-              <div className={styles.formRow}>
-                <span className={styles.formLabel}>Min element size</span>
-                <input
-                  className={styles.formInput}
-                  type="number"
-                  min={0}
-                  max={500}
-                  step={0.5}
-                  value={minElementSize}
-                  disabled={isMeshing}
-                  onChange={(e) =>
-                    setMinElementSize(Math.max(0, Number(e.target.value)))
-                  }
-                />
-                <span className={styles.toleranceUnit}>mm</span>
-              </div>
-              <div className={styles.formRow}>
-                <span className={styles.formLabel}>Element order</span>
-                <select
-                  className={styles.formSelect}
-                  value={elementOrder}
-                  onChange={(e) => setElementOrder(Number(e.target.value))}
-                >
-                  <option value={1}>Linear (1st order)</option>
-                  <option value={2}>Quadratic (2nd order)</option>
-                </select>
-              </div>
-              <button
-                className={hasMesh ? styles.outlineBtn : styles.meshVolBtn}
-                disabled={isMeshing}
-                onClick={handleVolMesh}
-              >
-                {isMeshing
-                  ? "Meshing…"
-                  : hasMesh
-                    ? "⟳ Re-mesh STEP volume"
-                    : "▶  Mesh STEP volume"}
-              </button>
-            </>
-          ) : (
-            !hasMesh && (
-              <div className={styles.empty}>
-                No mesh — import a STEP file to mesh.
-              </div>
-            )
-          )}
-
-          {/* VSCode-style collapsable log panel */}
-          {showLogs && (
-            <div className={styles.logSection}>
-              <button
-                className={styles.logHeader}
-                onClick={() => setLogsOpen((v) => !v)}
-              >
-                <span
-                  className={`${styles.logChevron} ${logsOpen ? styles.logChevronOpen : ""}`}
-                >
-                  ▶
-                </span>
-                <span>LOGS</span>
-                {isMeshing && <span className={styles.logSpinner}>●</span>}
-                {logs.length > 0 && (
-                  <span className={styles.logBadge}>{logs.length}</span>
-                )}
-              </button>
-              {logsOpen && (
-                <div className={styles.logBody}>
-                  {logs.length === 0 ? (
-                    <div className={styles.logEmpty}>Waiting…</div>
-                  ) : (
-                    logs.map((line, i) => (
-                      <div
-                        key={i}
-                        className={`${styles.logLine} ${i === logs.length - 1 ? styles.logLineLast : ""}`}
-                      >
-                        {line}
-                      </div>
-                    ))
-                  )}
-                  <div ref={logsEndRef} />
-                </div>
-              )}
-            </div>
-          )}
-        </>
-
-        {/* ── Material ──────────────────────────────────────────── */}
-        {/* Single-part models carry exactly one material, applied to the whole
-            part (issue #275). Per-body materials arrive with multibody support
-            (issue #317), so materials cannot be added here; delete is offered
-            only to trim legacy analyses saved with several. */}
-        <>
-          <div className={styles.sectionLabel}>Material</div>
-          {materials.length === 0 && (
-            <div className={styles.empty}>No material</div>
-          )}
-          {materials.length > 1 && (
-            <div
-              className={styles.errorBanner}
-              data-testid="material-count-warning"
-            >
-              <span>
-                Only one material is supported — it is applied to the whole
-                part. Remove the extra materials before solving.
-              </span>
-            </div>
-          )}
-          {materials.map((m) => (
-            <div key={m.id}>
-              <div className={styles.treeItem}>
-                <div className={styles.treeItemBody}>
-                  <div className={styles.treeItemName}>{m.name}</div>
-                  <div className={styles.treeItemDetail}>
-                    E = {fmt(m.young, 3)} MPa · ν = {m.poisson}
-                  </div>
-                </div>
-                <div className={styles.treeItemActions}>
-                  <button
-                    className={styles.iconBtn}
-                    title="Edit material"
-                    onClick={() => setEditingMatId(m.id)}
-                  >
-                    ✎
-                  </button>
-                  {materials.length > 1 && (
-                    <button
-                      className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                      title="Remove material"
-                      onClick={() => deleteMaterial(m.id)}
-                    >
-                      ×
-                    </button>
-                  )}
-                </div>
-              </div>
-              {editingMatId === m.id && (
-                <MaterialForm
-                  mat={m}
-                  onSave={(v) => {
-                    updateMaterial(m.id, v);
-                    setEditingMatId(null);
-                  }}
-                  onCancel={() => setEditingMatId(null)}
-                />
-              )}
-            </div>
-          ))}
-          {materials.length === 1 && (
-            <div className={styles.empty}>Applied to the whole part.</div>
-          )}
-        </>
-      </div>
-    </div>
-  );
-}
-
-// ── Constraints mode ──────────────────────────────────────────────────────────
-
-const DOF_LABELS = ["Ux", "Uy", "Uz", "Rx", "Ry", "Rz"];
-const FORCE_LABELS = ["Fx", "Fy", "Fz"];
-const MOMENT_LABELS = ["Mx", "My", "Mz"];
-
-// One-line summary of a load group for the group list: pressure magnitude, or
-// the non-zero force/moment components (e.g. "Fx = 100, Fz = -50 N").
-function loadGroupMeta(g: NamedLoadGroup): string {
-  const kind = loadKind(g);
-  if (kind === "pressure") return `p = ${fmt(g.totalForce)} MPa`;
-  const labels = kind === "moment" ? ["Mx", "My", "Mz"] : ["Fx", "Fy", "Fz"];
-  const unit = kind === "moment" ? "N·mm" : "N";
-  const parts = loadComponents(g)
-    .map((v, i) => (v !== 0 ? `${labels[i]} = ${fmt(v)}` : null))
-    .filter((p): p is string => p !== null);
-  return `${parts.join(", ")} ${unit}`;
-}
-
-// Inline editor for a BC group's constrained DOFs and prescribed value —
-// opened by the group's ✎ button (issue #258: the values, not just the faces,
-// must be editable after creation).
-function BcValueForm({
-  group,
-  onSave,
-  onCancel,
-}: {
-  group: NamedBcGroup;
-  onSave(dofs: number[], value: number): void;
-  onCancel(): void;
-}) {
-  const [checkedDofs, setCheckedDofs] = useState<boolean[]>(
-    DOF_LABELS.map((_, i) => group.dofs.includes(i)),
-  );
-  const [value, setValue] = useState(String(group.value));
-  const [error, setError] = useState<string | null>(null);
-
-  function handleSave() {
-    // A prescribed displacement of 0 is physically valid (fixed support), so
-    // only reject non-finite input — never silently coerce "abc" to 0.
-    const parsedValue = parseFloat(value);
-    if (!isFinite(parsedValue)) {
-      setError("Prescribed displacement must be a finite number");
-      return;
-    }
-    const dofs = checkedDofs.map((c, i) => (c ? i : -1)).filter((i) => i >= 0);
-    if (dofs.length === 0) {
-      setError("Constrain at least one DOF");
-      return;
-    }
-    onSave(dofs, parsedValue);
-  }
-
-  return (
-    <div className={styles.inlineForm} data-testid="bc-edit-form">
-      {error && (
-        <div className={styles.errorBanner} data-testid="bc-edit-error">
-          <span>{error}</span>
-          <button onClick={() => setError(null)}>×</button>
-        </div>
-      )}
-      <div className={styles.dofGrid}>
-        {/* Solid (H1 displacement) elements have only translational DOFs —
-        Ux, Uy, Uz. Rotational constraints carry no stiffness and are not
-        offered. */}
-        {DOF_LABELS.slice(0, 3).map((d, i) => (
-          <label key={d} className={styles.dofCheck}>
-            <input
-              type="checkbox"
-              checked={checkedDofs[i]}
-              onChange={() =>
-                setCheckedDofs((p) => p.map((v, j) => (j === i ? !v : v)))
-              }
-            />
-            {d}
-          </label>
-        ))}
-      </div>
-      <div className={styles.formRow}>
-        <span className={styles.formLabel}>Value</span>
-        <input
-          className={styles.formInput}
-          type="number"
-          value={value}
-          step="0.001"
-          onChange={(e) => setValue(e.target.value)}
-        />
-      </div>
-      <div className={styles.formBtns}>
-        <button className={styles.cancelBtn} onClick={onCancel}>
-          Cancel
-        </button>
-        <button className={styles.primaryBtn} onClick={handleSave}>
-          Save
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// Inline editor for a load group's kind and values — componentwise force /
-// moment vector or pressure magnitude — opened by the group's ✎ button
-// (issue #258).
-function LoadValueForm({
-  group,
-  onSave,
-  onCancel,
-}: {
-  group: NamedLoadGroup;
-  onSave(
-    kind: LoadKind,
-    totalForce: number,
-    components?: [number, number, number],
-  ): void;
-  onCancel(): void;
-}) {
-  const initialKind = loadKind(group);
-  const initialVec = loadComponents(group);
-  const [kindSel, setKindSel] = useState<LoadKind>(initialKind);
-  const [vec, setVec] = useState<[string, string, string]>([
-    String(initialVec[0]),
-    String(initialVec[1]),
-    String(initialVec[2]),
-  ]);
-  const [pressureVal, setPressureVal] = useState(
-    initialKind === "pressure" ? String(group.totalForce) : "10",
-  );
-  const [error, setError] = useState<string | null>(null);
-
-  function handleSave() {
-    if (kindSel === "pressure") {
-      // A zero pressure is a no-op load: it contributes nothing to the RHS and
-      // the solver returns a plausible-looking field with the input silently
-      // discarded. Reject it (and non-finite input) instead of coercing to 0.
-      const pressure = parseFloat(pressureVal);
-      if (!isFinite(pressure) || pressure === 0) {
-        setError("Pressure must be a non-zero finite number");
-        return;
-      }
-      onSave("pressure", pressure);
-      return;
-    }
-    // Force / moment, prescribed componentwise. Reject non-finite components
-    // and the all-zero vector (a no-op load that would be silently discarded).
-    const noun = kindSel === "moment" ? "moment" : "force";
-    const parsed = vec.map((comp) => parseFloat(comp));
-    if (parsed.some((comp) => !isFinite(comp))) {
-      setError(`Each ${noun} component must be a finite number`);
-      return;
-    }
-    if (parsed.every((comp) => comp === 0)) {
-      setError(`Specify a non-zero ${noun} component`);
-      return;
-    }
-    onSave(kindSel, 0, [parsed[0], parsed[1], parsed[2]]);
-  }
-
-  const labels = kindSel === "moment" ? MOMENT_LABELS : FORCE_LABELS;
-  const unit = kindSel === "moment" ? "N·mm" : "N";
-
-  return (
-    <div className={styles.inlineForm} data-testid="load-edit-form">
-      {error && (
-        <div className={styles.errorBanner} data-testid="load-edit-error">
-          <span>{error}</span>
-          <button onClick={() => setError(null)}>×</button>
-        </div>
-      )}
-      <div className={styles.formRow}>
-        <span className={styles.formLabel}>Type</span>
-        <select
-          className={styles.formSelect}
-          value={kindSel}
-          onChange={(e) => setKindSel(e.target.value as LoadKind)}
-        >
-          <option value="force">Force</option>
-          <option value="moment">Moment</option>
-          <option value="pressure">Pressure</option>
-        </select>
-      </div>
-      {kindSel === "pressure" ? (
-        <div className={styles.formRow}>
-          <span className={styles.formLabel}>Pressure (MPa)</span>
-          <input
-            className={styles.formInput}
-            type="number"
-            value={pressureVal}
-            step="1"
-            onChange={(e) => setPressureVal(e.target.value)}
-          />
-        </div>
-      ) : (
-        labels.map((label, i) => (
-          <div className={styles.formRow} key={label}>
-            <span className={styles.formLabel}>
-              {label} ({unit})
-            </span>
-            <input
-              className={styles.formInput}
-              type="number"
-              value={vec[i]}
-              step="100"
-              onChange={(e) => {
-                const value = e.target.value;
-                setVec((p) => {
-                  const next = [...p] as [string, string, string];
-                  next[i] = value;
-                  return next;
-                });
-              }}
-            />
-          </div>
-        ))
-      )}
-      <div className={styles.formBtns}>
-        <button className={styles.cancelBtn} onClick={onCancel}>
-          Cancel
-        </button>
-        <button className={styles.primaryBtn} onClick={handleSave}>
-          Save
-        </button>
-      </div>
-    </div>
-  );
-}
-
-function ConstraintsPanel() {
-  const nodes = useModelStore((s) => s.nodes);
-  const bcGroups = useModelStore((s) => s.bcGroups);
-  const loadGroups = useModelStore((s) => s.loadGroups);
-  const pickMode = useModelStore((s) => s.pickMode);
-  const pickTargetGroupId = useModelStore((s) => s.pickTargetGroupId);
-  const setPickMode = useModelStore((s) => s.setPickMode);
-  const selectedFace = useModelStore((s) => s.selectedFace);
-  const setSelectedFace = useModelStore((s) => s.setSelectedFace);
-  const pendingFaces = useModelStore((s) => s.pendingFaces);
-  const setPendingFaces = useModelStore((s) => s.setPendingFaces);
-  const createBcGroup = useModelStore((s) => s.createBcGroup);
-  const addFaceToBcGroup = useModelStore((s) => s.addFaceToBcGroup);
-  const removeFaceFromBcGroup = useModelStore((s) => s.removeFaceFromBcGroup);
-  const deleteBcGroup = useModelStore((s) => s.deleteBcGroup);
-  const createLoadGroup = useModelStore((s) => s.createLoadGroup);
-  const addFaceToLoadGroup = useModelStore((s) => s.addFaceToLoadGroup);
-  const removeFaceFromLoadGroup = useModelStore(
-    (s) => s.removeFaceFromLoadGroup,
-  );
-  const deleteLoadGroup = useModelStore((s) => s.deleteLoadGroup);
-  const updateBcGroup = useModelStore((s) => s.updateBcGroup);
-  const updateLoadGroup = useModelStore((s) => s.updateLoadGroup);
-
-  // Group whose values are being edited inline (✎), if any.
-  const [editingBcId, setEditingBcId] = useState<number | null>(null);
-  const [editingLoadId, setEditingLoadId] = useState<number | null>(null);
-
-  const [checkedDofs, setCheckedDofs] = useState([
-    true,
-    true,
-    true,
-    false,
-    false,
-    false,
-  ]);
-  // A load is defined in two steps (issues #219, #190): pick the kind, then
-  // prescribe each component on its own — like the displacement BC above. Force
-  // and moment carry a [x,y,z] vector; pressure is a single scalar.
-  const [loadKindSel, setLoadKindSel] = useState<LoadKind>("force");
-  const [forceVec, setForceVec] = useState<[string, string, string]>([
-    "0",
-    "-10000",
-    "0",
-  ]);
-  const [momentVec, setMomentVec] = useState<[string, string, string]>([
-    "0",
-    "0",
-    "1000",
-  ]);
-  const [pressureVal, setPressureVal] = useState("10");
-  const [bcValue, setBcValue] = useState("0");
-  const [error, setError] = useState<string | null>(null);
-
-  // Boundary conditions and loads reference mesh nodes, so they can only be
-  // defined once a volume mesh exists.
-  const meshOk = nodes.length > 0;
-
-  const targetBcGroup =
-    pickTargetGroupId !== null
-      ? (bcGroups.find((g) => g.id === pickTargetGroupId) ?? null)
-      : null;
-  const targetLoadGroup =
-    pickTargetGroupId !== null
-      ? (loadGroups.find((g) => g.id === pickTargetGroupId) ?? null)
-      : null;
-
-  const allPickedFaces = selectedFace
-    ? [...pendingFaces, selectedFace]
-    : pendingFaces;
-
-  function removePickedFace(index: number) {
-    if (index < pendingFaces.length) {
-      setPendingFaces(pendingFaces.filter((_, i) => i !== index));
-    } else {
-      setSelectedFace(null);
-    }
-  }
-
-  function cancelPick() {
-    setError(null);
-    setPickMode(null);
-    setSelectedFace(null);
-    setPendingFaces([]);
-  }
-
-  function applyBc() {
-    if (allPickedFaces.length === 0) return;
-    const existingCount = targetBcGroup?.faces.length ?? 0;
-    const faceEntries = allPickedFaces.map((f, i) => ({
-      label: `Face ${existingCount + i + 1}`,
-      nodeIds: f.nodeIds,
-    }));
-    if (targetBcGroup) {
-      for (const fe of faceEntries) addFaceToBcGroup(targetBcGroup.id, fe);
-    } else {
-      // A prescribed displacement of 0 is physically valid (fixed support), so
-      // only reject non-finite input — never silently coerce "abc" to 0.
-      const value = parseFloat(bcValue);
-      if (!isFinite(value)) {
-        setError("Prescribed displacement must be a finite number");
-        return;
-      }
-      const dofs = checkedDofs
-        .map((c, i) => (c ? i : -1))
-        .filter((i) => i >= 0);
-      createBcGroup(faceEntries, dofs, value);
-    }
-    setError(null);
-    setPickMode(null);
-    setSelectedFace(null);
-    setPendingFaces([]);
-  }
-
-  function applyLoad() {
-    if (allPickedFaces.length === 0) return;
-    const existingCount = targetLoadGroup?.faces.length ?? 0;
-    const faceEntries = allPickedFaces.map((f, i) => ({
-      label: `Face ${existingCount + i + 1}`,
-      nodeIds: f.nodeIds,
-    }));
-    if (targetLoadGroup) {
-      for (const fe of faceEntries) addFaceToLoadGroup(targetLoadGroup.id, fe);
-    } else if (loadKindSel === "pressure") {
-      // A zero pressure is a no-op load: it contributes nothing to the RHS and
-      // the solver returns a plausible-looking field with the input silently
-      // discarded. Reject it (and non-finite input) instead of coercing to 0.
-      const p = parseFloat(pressureVal);
-      if (!isFinite(p) || p === 0) {
-        setError("Pressure must be a non-zero finite number");
-        return;
-      }
-      createLoadGroup(faceEntries, 0, p, "pressure");
-    } else {
-      // Force / moment, prescribed componentwise. Reject non-finite components
-      // and the all-zero vector (a no-op load that would be silently discarded).
-      const noun = loadKindSel === "moment" ? "moment" : "force";
-      const parsed = (loadKindSel === "moment" ? momentVec : forceVec).map(
-        (comp) => parseFloat(comp),
-      );
-      if (parsed.some((comp) => !isFinite(comp))) {
-        setError(`Each ${noun} component must be a finite number`);
-        return;
-      }
-      if (parsed.every((comp) => comp === 0)) {
-        setError(`Specify a non-zero ${noun} component`);
-        return;
-      }
-      createLoadGroup(faceEntries, 0, 0, loadKindSel, [
-        parsed[0],
-        parsed[1],
-        parsed[2],
-      ]);
-    }
-    setError(null);
-    setPickMode(null);
-    setSelectedFace(null);
-    setPendingFaces([]);
-  }
-
-  return (
-    <div className={styles.panel}>
-      <div className={styles.tabContent}>
-        {error && (
-          <div className={styles.errorBanner} data-testid="constraints-error">
-            <span>{error}</span>
-            <button onClick={() => setError(null)}>×</button>
-          </div>
-        )}
-        {!meshOk && (
-          <div className={styles.empty} data-testid="no-mesh-hint">
-            Generate a mesh before adding boundary conditions.
-          </div>
-        )}
-
-        {meshOk && (
-          <>
-            {/* ── BC section ────────────────────────────────────── */}
-            <div className={styles.sectionLabel}>Fixed displacement</div>
-
-            {pickMode !== "bc" && (
-              <button
-                className={styles.pickBtn}
-                onClick={() => setPickMode("bc", null)}
-              >
-                + Add BC
-              </button>
-            )}
-
-            {pickMode === "bc" && (
-              <div className={styles.pickPanel}>
-                <div className={styles.pickPanelHeader}>
-                  <span className={styles.pickPanelTitle}>
-                    {targetBcGroup
-                      ? `Add face to ${targetBcGroup.name}`
-                      : "New BC"}
-                  </span>
-                  <button
-                    className={styles.iconBtn}
-                    onClick={cancelPick}
-                    title="Cancel"
-                  >
-                    ✕
-                  </button>
-                </div>
-
-                {allPickedFaces.length === 0 ? (
-                  <div className={styles.pickHint}>
-                    Click a face in the 3D viewport
-                  </div>
-                ) : (
-                  <div>
-                    {allPickedFaces.map((f, i) => (
-                      <div key={i} className={styles.bcFaceRow}>
-                        <span className={styles.bcFaceName}>{f.label}</span>
-                        <button
-                          className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                          title="Remove face"
-                          onClick={() => removePickedFace(i)}
-                        >
-                          ×
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                )}
-
-                {allPickedFaces.length > 0 && !targetBcGroup && (
-                  <>
-                    <div className={styles.dofGrid}>
-                      {/* Solid (H1 displacement) elements have only translational
-                      DOFs — Ux, Uy, Uz. Rotational constraints carry no
-                      stiffness and are not offered. */}
-                      {DOF_LABELS.slice(0, 3).map((d, i) => (
-                        <label key={d} className={styles.dofCheck}>
-                          <input
-                            type="checkbox"
-                            checked={checkedDofs[i]}
-                            onChange={() =>
-                              setCheckedDofs((p) =>
-                                p.map((v, j) => (j === i ? !v : v)),
-                              )
-                            }
-                          />
-                          {d}
-                        </label>
-                      ))}
-                    </div>
-                    <div className={styles.formRow}>
-                      <span className={styles.formLabel}>Value</span>
-                      <input
-                        className={styles.formInput}
-                        type="number"
-                        value={bcValue}
-                        step="0.001"
-                        onChange={(e) => setBcValue(e.target.value)}
-                      />
-                    </div>
-                    <button className={styles.primaryBtn} onClick={applyBc}>
-                      Apply BC
-                    </button>
-                  </>
-                )}
-
-                {allPickedFaces.length > 0 && targetBcGroup && (
-                  <button className={styles.primaryBtn} onClick={applyBc}>
-                    {allPickedFaces.length === 1
-                      ? "Add Face"
-                      : `Add ${allPickedFaces.length} Faces`}
-                  </button>
-                )}
-              </div>
-            )}
-
-            {/* BC group list */}
-            {bcGroups.map((g) => (
-              <div key={g.id} className={styles.bcGroup}>
-                <div className={styles.bcGroupHeader}>
-                  <span className={styles.bcDot} />
-                  <span className={styles.bcGroupName}>{g.name}</span>
-                  <span className={styles.bcGroupMeta}>
-                    {g.dofs.map((d) => DOF_LABELS[d]).join(", ")} = {g.value}
-                  </span>
-                  <div className={styles.treeItemActions}>
-                    <button
-                      className={styles.iconBtn}
-                      title="Add face"
-                      onClick={() => {
-                        setPickMode("bc", g.id);
-                        setSelectedFace(null);
-                        setPendingFaces([]);
-                      }}
-                    >
-                      +
-                    </button>
-                    <button
-                      className={styles.iconBtn}
-                      title="Edit BC"
-                      onClick={() =>
-                        setEditingBcId(editingBcId === g.id ? null : g.id)
-                      }
-                    >
-                      ✎
-                    </button>
-                    <button
-                      className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                      title="Delete BC"
-                      onClick={() => deleteBcGroup(g.id)}
-                    >
-                      ✕
-                    </button>
-                  </div>
-                </div>
-                {editingBcId === g.id && (
-                  <BcValueForm
-                    group={g}
-                    onSave={(dofs, value) => {
-                      updateBcGroup(g.id, dofs, value);
-                      setEditingBcId(null);
-                    }}
-                    onCancel={() => setEditingBcId(null)}
-                  />
-                )}
-                {g.faces.map((f) => (
-                  <div key={f.id} className={styles.bcFaceRow}>
-                    <span className={styles.bcFaceIndent}>└</span>
-                    <span className={styles.bcFaceName}>{f.label}</span>
-                    <button
-                      className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                      title="Remove face"
-                      onClick={() => removeFaceFromBcGroup(g.id, f.id)}
-                    >
-                      ×
-                    </button>
-                  </div>
-                ))}
-              </div>
-            ))}
-
-            {/* ── Load section ───────────────────────────────────── */}
-            <div className={styles.sectionLabel} style={{ marginTop: 16 }}>
-              Applied loads
-            </div>
-
-            {pickMode !== "load" && (
-              <button
-                className={styles.pickBtn}
-                onClick={() => setPickMode("load", null)}
-              >
-                + Add Load
-              </button>
-            )}
-
-            {pickMode === "load" && (
-              <div className={styles.pickPanel}>
-                <div className={styles.pickPanelHeader}>
-                  <span className={styles.pickPanelTitle}>
-                    {targetLoadGroup
-                      ? `Add face to ${targetLoadGroup.name}`
-                      : "New Load"}
-                  </span>
-                  <button
-                    className={styles.iconBtn}
-                    onClick={cancelPick}
-                    title="Cancel"
-                  >
-                    ✕
-                  </button>
-                </div>
-
-                {allPickedFaces.length === 0 ? (
-                  <div className={styles.pickHint}>
-                    Click a face in the 3D viewport
-                  </div>
-                ) : (
-                  <div>
-                    {allPickedFaces.map((f, i) => (
-                      <div key={i} className={styles.bcFaceRow}>
-                        <span className={styles.bcFaceName}>{f.label}</span>
-                        <button
-                          className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                          title="Remove face"
-                          onClick={() => removePickedFace(i)}
-                        >
-                          ×
-                        </button>
-                      </div>
-                    ))}
-                  </div>
-                )}
-
-                {allPickedFaces.length > 0 && !targetLoadGroup && (
-                  <>
-                    {/* Step 1 — pick the load kind. */}
-                    <div className={styles.formRow}>
-                      <span className={styles.formLabel}>Type</span>
-                      <select
-                        className={styles.formSelect}
-                        value={loadKindSel}
-                        onChange={(e) =>
-                          setLoadKindSel(e.target.value as LoadKind)
-                        }
-                      >
-                        <option value="force">Force</option>
-                        <option value="moment">Moment</option>
-                        <option value="pressure">Pressure</option>
-                      </select>
-                    </div>
-
-                    {/* Step 2 — prescribe each component on its own. */}
-                    {loadKindSel === "pressure" ? (
-                      <div className={styles.formRow}>
-                        <span className={styles.formLabel}>Pressure (MPa)</span>
-                        <input
-                          className={styles.formInput}
-                          type="number"
-                          value={pressureVal}
-                          step="1"
-                          onChange={(e) => setPressureVal(e.target.value)}
-                        />
-                      </div>
-                    ) : (
-                      (loadKindSel === "moment"
-                        ? MOMENT_LABELS
-                        : FORCE_LABELS
-                      ).map((label, i) => {
-                        const vec =
-                          loadKindSel === "moment" ? momentVec : forceVec;
-                        const setVec =
-                          loadKindSel === "moment" ? setMomentVec : setForceVec;
-                        const unit = loadKindSel === "moment" ? "N·mm" : "N";
-                        return (
-                          <div className={styles.formRow} key={label}>
-                            <span className={styles.formLabel}>
-                              {label} ({unit})
-                            </span>
-                            <input
-                              className={styles.formInput}
-                              type="number"
-                              value={vec[i]}
-                              step="100"
-                              onChange={(e) => {
-                                const value = e.target.value;
-                                setVec((p) => {
-                                  const next = [...p] as [
-                                    string,
-                                    string,
-                                    string,
-                                  ];
-                                  next[i] = value;
-                                  return next;
-                                });
-                              }}
-                            />
-                          </div>
-                        );
-                      })
-                    )}
-                    <div className={styles.pickNote}>
-                      {loadKindSel === "pressure"
-                        ? "applied as p·n̂ over each face (work-equivalent)"
-                        : loadKindSel === "force"
-                          ? "applied as a work-equivalent surface traction"
-                          : "distributed as equivalent nodal forces"}
-                    </div>
-                    <button className={styles.loadBtn} onClick={applyLoad}>
-                      Apply Load
-                    </button>
-                  </>
-                )}
-
-                {allPickedFaces.length > 0 && targetLoadGroup && (
-                  <button className={styles.loadBtn} onClick={applyLoad}>
-                    {allPickedFaces.length === 1
-                      ? "Add Face"
-                      : `Add ${allPickedFaces.length} Faces`}
-                  </button>
-                )}
-              </div>
-            )}
-
-            {/* Load group list */}
-            {loadGroups.map((g) => (
-              <div key={g.id} className={styles.bcGroup}>
-                <div className={styles.bcGroupHeader}>
-                  <span className={styles.loadDot} />
-                  <span className={styles.bcGroupName}>{g.name}</span>
-                  <span className={styles.bcGroupMeta}>{loadGroupMeta(g)}</span>
-                  <div className={styles.treeItemActions}>
-                    <button
-                      className={styles.iconBtn}
-                      title="Add face"
-                      onClick={() => {
-                        setPickMode("load", g.id);
-                        setSelectedFace(null);
-                        setPendingFaces([]);
-                      }}
-                    >
-                      +
-                    </button>
-                    <button
-                      className={styles.iconBtn}
-                      title="Edit load"
-                      onClick={() =>
-                        setEditingLoadId(editingLoadId === g.id ? null : g.id)
-                      }
-                    >
-                      ✎
-                    </button>
-                    <button
-                      className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                      title="Delete Load"
-                      onClick={() => deleteLoadGroup(g.id)}
-                    >
-                      ✕
-                    </button>
-                  </div>
-                </div>
-                {editingLoadId === g.id && (
-                  <LoadValueForm
-                    group={g}
-                    onSave={(kind, totalForce, components) => {
-                      updateLoadGroup(g.id, kind, totalForce, components);
-                      setEditingLoadId(null);
-                    }}
-                    onCancel={() => setEditingLoadId(null)}
-                  />
-                )}
-                {g.faces.map((f) => (
-                  <div key={f.id} className={styles.bcFaceRow}>
-                    <span className={styles.bcFaceIndent}>└</span>
-                    <span className={styles.bcFaceName}>{f.label}</span>
-                    <button
-                      className={`${styles.iconBtn} ${styles.iconBtnDanger}`}
-                      title="Remove face"
-                      onClick={() => removeFaceFromLoadGroup(g.id, f.id)}
-                    >
-                      ×
-                    </button>
-                  </div>
-                ))}
-              </div>
-            ))}
-          </>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ── Solve mode ────────────────────────────────────────────────────────────────
-
-function SolvePanel() {
-  const nodes = useModelStore((s) => s.nodes);
-  const elements = useModelStore((s) => s.elements);
-  const materials = useModelStore((s) => s.materials);
-  const properties = useModelStore((s) => s.properties);
-  const constraints = useModelStore((s) => s.constraints);
-  const loads = useModelStore((s) => s.loads);
-  const surfaceLoads = useModelStore((s) => s.surfaceLoads);
-  const isRunning = useModelStore((s) => s.isRunning);
-  const setRunning = useModelStore((s) => s.setRunning);
-  const setResult = useModelStore((s) => s.setResult);
-  const setMode = useModelStore((s) => s.setMode);
-  const elementOrder = useModelStore((s) => s.elementOrder);
-  const [error, setError] = useState<string | null>(null);
-
-  const meshOk = nodes.length > 0;
-  const matOk = materials.length > 0;
-  const bcOk = constraints.length > 0;
-  // Force/pressure loads now reach the solver as surface tractions, moments as
-  // nodal forces — either kind makes the model loaded.
-  const loadOk = loads.length > 0 || surfaceLoads.length > 0;
-  // A non-zero prescribed displacement drives the deformation on its own, so the
-  // model is solvable without an applied load. Without either, the solve returns
-  // the trivial zero field, so at least one driving action is still required.
-  const prescribedOk = constraints.some((c) => (c.prescribedValue ?? 0) !== 0);
-  const drivingOk = loadOk || prescribedOk;
-  const allOk = meshOk && matOk && bcOk && drivingOk;
-
-  function handleSolve() {
-    setRunning(true);
-    sendToWorker<{ displacements: number[]; vonMises: number[] }>("solve", {
-      nodes,
-      elements,
-      materials,
-      properties,
-      constraints,
-      loads,
-      surfaceLoads,
-      elementOrder,
-    })
-      .then(({ displacements, vonMises }) => {
-        setResult({
-          displacements: new Float64Array(displacements),
-          vonMises: vonMises ? new Float64Array(vonMises) : undefined,
-        });
-        setMode("results");
-      })
-      .catch((err) => {
-        console.error("[solve] solver failed:", err.message);
-        setError(`Solver error: ${err.message}`);
-      })
-      .finally(() => setRunning(false));
-  }
-
-  // Expose for Playwright E2E tests — allows bypassing the button's disabled-state
-  // timing uncertainty in CI without requiring UI interaction.
-  useEffect(() => {
-    (
-      window as Window & { __kofemTriggerSolve?: () => void }
-    ).__kofemTriggerSolve = handleSolve;
-  });
-
-  const checks: [boolean, string][] = [
-    [
-      meshOk,
-      `Mesh ready · ${nodes.length} nodes · ${elements.length} elements`,
-    ],
-    [
-      matOk,
-      materials.length > 0
-        ? `Material assigned · ${materials[0].name} · E=${fmt(materials[0].young, 3)} MPa`
-        : "No material assigned",
-    ],
-    [
-      bcOk,
-      bcOk
-        ? `BCs applied · ${new Set(constraints.map((c) => c.nodeId)).size} nodes fixed`
-        : "No boundary conditions",
-    ],
-    [
-      drivingOk,
-      loadOk
-        ? `Loads applied · ${loads.length} load DOFs`
-        : prescribedOk
-          ? "Prescribed displacement drives the model"
-          : "No loads or prescribed displacement",
-    ],
-  ];
-
-  return (
-    <div className={styles.panel}>
-      <div className={styles.tabContent}>
-        {error && (
-          <div className={styles.errorBanner}>
-            <span>{error}</span>
-            <button onClick={() => setError(null)}>×</button>
-          </div>
-        )}
-        <div className={styles.sectionLabel}>Pre-flight check</div>
-        {checks.map(([ok, label], i) => (
-          <div key={i} className={styles.checkRow}>
-            <span className={ok ? styles.checkOk : styles.checkFail}>
-              {ok ? "✓" : "✗"}
-            </span>
-            <span className={styles.checkLabel}>{label}</span>
-          </div>
-        ))}
-
-        <div className={styles.sectionLabel} style={{ marginTop: 16 }}>
-          Solver settings
-        </div>
-        <div className={styles.statRow}>
-          <span className={styles.statKey}>Step</span>
-          <span className={styles.statVal}>Static · Step-1</span>
-        </div>
-        <div className={styles.statRow}>
-          <span className={styles.statKey}>Solver</span>
-          <span className={styles.statVal}>Direct</span>
-        </div>
-        <div className={styles.statRow}>
-          <span className={styles.statKey}>Output</span>
-          <span className={styles.statVal}>U, S, RF</span>
-        </div>
-        {/* The element order is set with the mesh controls on the Geometry
-            tab (issue #286) — echoed here read-only so the pre-solve summary
-            stays complete. */}
-        <div className={styles.statRow}>
-          <span className={styles.statKey}>Element order</span>
-          <span className={styles.statVal}>
-            {elementOrder === 2
-              ? "Quadratic (2nd order)"
-              : "Linear (1st order)"}
-          </span>
-        </div>
-
-        <button
-          className={styles.solveBtn}
-          disabled={!allOk || isRunning}
-          onClick={handleSolve}
-        >
-          {isRunning ? "Solving…" : "▶  Run static solve"}
-        </button>
-      </div>
-    </div>
-  );
-}
-
-// ── Results mode ──────────────────────────────────────────────────────────────
-
-function ResultsPanel() {
-  const result = useModelStore((s) => s.result);
-  const resultType = useModelStore((s) => s.resultType);
-  const setResultType = useModelStore((s) => s.setResultType);
-  const deformScale = useModelStore((s) => s.deformScale);
-  const setDeformScale = useModelStore((s) => s.setDeformScale);
-  const nodes = useModelStore((s) => s.nodes);
-  const elements = useModelStore((s) => s.elements);
-
-  if (!result) {
-    return (
-      <div className={styles.panel}>
-        <div className={styles.tabContent}>
-          <div className={styles.empty}>No results — run the solver first</div>
-        </div>
-      </div>
-    );
-  }
-
-  // Min/max of the selected scalar field over all nodes — the same field and
-  // node averaging used for the viewport coloring and colorbar legend.
-  const stats = computeResultRange(result, resultType, nodes, elements);
-
-  const fieldSymbol = resultFieldSymbol(resultType);
-  const unit = resultUnit(resultType);
-
-  return (
-    <div className={styles.panel}>
-      <div className={styles.tabContent}>
-        <div className={styles.sectionLabel}>Field</div>
-        <select
-          className={styles.formSelect}
-          style={{ marginBottom: 12 }}
-          value={resultType}
-          onChange={(e) => setResultType(e.target.value as ResultType)}
-        >
-          {RESULT_TYPES.map((t) => (
-            <option key={t} value={t}>
-              {t}
-            </option>
-          ))}
-        </select>
-
-        <div className={styles.sectionLabel}>Deformation scale</div>
-        <div
-          style={{
-            display: "flex",
-            alignItems: "center",
-            gap: 8,
-            marginBottom: 12,
-          }}
-        >
-          <input
-            type="range"
-            min={0}
-            max={3}
-            step={0.05}
-            value={deformScale}
-            onChange={(e) => setDeformScale(parseFloat(e.target.value))}
-            style={{ flex: 1 }}
-            aria-label="Deformation scale"
-          />
-          <span
-            className={styles.statVal}
-            style={{ minWidth: 38, textAlign: "right" }}
-          >
-            {deformScale.toFixed(2)}×
-          </span>
-        </div>
-
-        <div className={styles.sectionLabel}>Result summary</div>
-        {stats ? (
-          <>
-            <div className={styles.statRow}>
-              <span className={styles.statKey}>Min {fieldSymbol}</span>
-              <span className={styles.statVal}>
-                {stats.min.toExponential(3)} {unit}
-              </span>
-            </div>
-            <div className={styles.statRow}>
-              <span className={styles.statKey}>Max {fieldSymbol}</span>
-              <span className={styles.statVal}>
-                {stats.max.toExponential(3)} {unit}
-              </span>
-            </div>
-          </>
-        ) : (
-          <div className={styles.empty}>
-            {resultType === "Von Mises stress"
-              ? "Von Mises data not available — re-run the solver"
-              : "No nodal data"}
-          </div>
-        )}
-      </div>
-    </div>
-  );
-}
-
-// ── Nav ───────────────────────────────────────────────────────────────────────
-
-const MODES: { id: AppMode; label: string }[] = [
-  { id: "geometry", label: "Geometry" },
-  { id: "constraints", label: "Constraints" },
-  { id: "solve", label: "Solve" },
-  { id: "results", label: "Results" },
-];
-
-function PanelNav() {
-  const mode = useModelStore((s) => s.mode);
-  const setMode = useModelStore((s) => s.setMode);
-  const nodes = useModelStore((s) => s.nodes);
-  const constraints = useModelStore((s) => s.constraints);
-  const loads = useModelStore((s) => s.loads);
-  const result = useModelStore((s) => s.result);
-
-  function isValid(m: AppMode): boolean {
-    if (m === "geometry") return nodes.length > 0;
-    if (m === "constraints") return constraints.length > 0 || loads.length > 0;
-    if (m === "solve" || m === "results") return result !== null;
-    return false;
-  }
-
-  return (
-    <nav className={styles.modeNav}>
-      {MODES.map(({ id, label }) => {
-        const active = id === mode;
-        const valid = isValid(id);
-        return (
-          <button
-            key={id}
-            className={`${styles.navTab} ${active ? styles.navTabActive : styles.navTabFuture}`}
-            onClick={() => setMode(id)}
-          >
-            {valid ? (
-              <span className={`${styles.navDot} ${styles.navDotDone}`}>
-                <svg viewBox="0 0 8 8" width="5" height="5">
-                  <path
-                    d="M1.5 4L3 5.5L6.5 2"
-                    stroke="currentColor"
-                    strokeWidth="1.8"
-                    fill="none"
-                    strokeLinecap="round"
-                  />
-                </svg>
-              </span>
-            ) : (
-              <span className={styles.navDot} />
-            )}
-            <span className={styles.navTabLabel}>{label}</span>
-          </button>
-        );
-      })}
-    </nav>
-  );
-}
-
-// ── Root ──────────────────────────────────────────────────────────────────────
-
+// Thin orchestrator: the per-mode panels own their state and worker calls
+// (see hooks/useGeometry, useMesh, useSolver) — this component only maps the
+// active mode to a panel.
 export function LeftPanel() {
   const mode = useModelStore((s) => s.mode);
   return (
     <aside className={styles.aside}>
       <PanelNav />
       {mode === "geometry" && <GeometryPanel />}
-      {mode === "constraints" && <ConstraintsPanel />}
+      {mode === "constraints" && <BoundaryConditionsPanel />}
       {mode === "solve" && <SolvePanel />}
       {mode === "results" && <ResultsPanel />}
     </aside>
